Move convertGithubNwoToDatabaseUrl to databases/github-databases/api.ts
