+++ extensions/ql-vscode/src/databases/database-fetcher.ts
@@ -33,11 +33,11 @@ import {
 } from "../config";
 import { showAndLogInformationMessage } from "../common/logging";
 import { AppOctokit } from "../common/octokit";
-import { getLanguageDisplayName } from "../common/query-language";
 import type { DatabaseOrigin } from "./local-databases/database-origin";
 import { createTimeoutSignal } from "../common/fetch-stream";
 import type { App } from "../common/app";
 import { findDirWithFile } from "../common/files";
+import { convertGithubNwoToDatabaseUrl } from "./github-databases/api";
 
 /**
  * Prompts a user to fetch a database from a remote location. Database is assumed to be an archive file.
@@ -585,95 +585,6 @@ async function checkForFailingResponse(
 
 function isFile(databaseUrl: string) {
   return Uri.parse(databaseUrl).scheme === "file";
-}
-
-export async function convertGithubNwoToDatabaseUrl(
-  nwo: string,
-  octokit: Octokit,
-  progress: ProgressCallback,
-  language?: string,
-): Promise<
-  | {
-      databaseUrl: string;
-      owner: string;
-      name: string;
-      databaseId: number;
-      databaseCreatedAt: string;
-      commitOid: string | null;
-    }
-  | undefined
-> {
-  try {
-    const [owner, repo] = nwo.split("/");
-
-    const response = await octokit.rest.codeScanning.listCodeqlDatabases({
-      owner,
-      repo,
-    });
-
-    const languages = response.data.map((db) => db.language);
-
-    if (!language || !languages.includes(language)) {
-      language = await promptForLanguage(languages, progress);
-      if (!language) {
-        return;
-      }
-    }
-
-    const databaseForLanguage = response.data.find(
-      (db) => db.language === language,
-    );
-    if (!databaseForLanguage) {
-      throw new Error(`No database found for language '${language}'`);
-    }
-
-    return {
-      databaseUrl: databaseForLanguage.url,
-      owner,
-      name: repo,
-      databaseId: databaseForLanguage.id,
-      databaseCreatedAt: databaseForLanguage.created_at,
-      commitOid: databaseForLanguage.commit_oid ?? null,
-    };
-  } catch (e) {
-    void extLogger.log(`Error: ${getErrorMessage(e)}`);
-    throw new Error(`Unable to get database for '${nwo}'`);
-  }
-}
-
-export async function promptForLanguage(
-  languages: string[],
-  progress: ProgressCallback | undefined,
-): Promise<string | undefined> {
-  progress?.({
-    message: "Choose language",
-    step: 2,
-    maxStep: 2,
-  });
-  if (!languages.length) {
-    throw new Error("No databases found");
-  }
-  if (languages.length === 1) {
-    return languages[0];
-  }
-
-  const items = languages
-    .map((language) => ({
-      label: getLanguageDisplayName(language),
-      description: language,
-      language,
-    }))
-    .sort((a, b) => a.label.localeCompare(b.label));
-
-  const selectedItem = await window.showQuickPick(items, {
-    placeHolder: "Select the database language to download:",
-    ignoreFocusOut: true,
-  });
-  if (!selectedItem) {
-    return undefined;
-  }
-
-  return selectedItem.language;
 }
 
 /**
